+++ packages/core/src/render3/index.ts
@@ -70,7 +70,6 @@ export {
 
   ɵɵgetCurrentView,
   ɵɵinjectAttribute,
-  ɵɵinjectPipeChangeDetectorRef,
 
   ɵɵlistener,
   ɵɵload,
@@ -199,7 +198,7 @@ export {
   ɵɵpureFunctionV,
 } from './pure_function';
 
-export {ɵɵtemplateRefExtractor} from './view_engine_compatibility_prebound';
+export {ɵɵtemplateRefExtractor, ɵɵinjectPipeChangeDetectorRef} from './view_engine_compatibility_prebound';
 
 export {ɵɵresolveWindow, ɵɵresolveDocument, ɵɵresolveBody} from './util/misc_utils';
